moving commands to hoprd
