+++ packages/hoprd/src/commands/cover-traffic.ts
