+++ packages/hoprd/src/commands/listConnected.ts
