+++ packages/hoprd/src/commands/listConnectors.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk'
-import { knownConnectors } from '../utils/knownConnectors'
+import { knownConnectors } from './utils/knownConnectors'
 import { AbstractCommand } from './abstractCommand'
 
 export default class ListConnectors extends AbstractCommand {
