+++ packages/hoprd/src/commands/multisend.ts
@@ -4,7 +4,7 @@ import { clearString } from '@hoprnet/hopr-utils'
 import { SendMessageBase } from './sendMessage'
 import readline from 'readline'
 import type PeerId from 'peer-id'
-import { getPeersIdsAsString, checkPeerIdInput, styleValue } from '../utils'
+import { getPeersIdsAsString, checkPeerIdInput, styleValue } from './utils'
 import { GlobalState, AutoCompleteResult, CommandResponse } from './abstractCommand'
 
 export class MultiSendMessage extends SendMessageBase {
