+++ packages/hoprd/src/commands/ping.ts
@@ -3,7 +3,7 @@ import type Hopr from '@hoprnet/hopr-core'
 import type PeerId from 'peer-id'
 import type { AutoCompleteResult } from './abstractCommand'
 import { AbstractCommand, GlobalState } from './abstractCommand'
-import { checkPeerIdInput, isBootstrapNode, getPeerIdsAndAliases, styleValue } from '../utils'
+import { checkPeerIdInput, isBootstrapNode, getPeerIdsAndAliases, styleValue } from './utils'
 
 export default class Ping extends AbstractCommand {
   constructor(public node: Hopr<HoprCoreConnector>) {
