+++ packages/hoprd/src/commands/printAddress.ts
@@ -2,7 +2,7 @@ import type HoprCoreConnector from '@hoprnet/hopr-core-connector-interface'
 import type Hopr from '@hoprnet/hopr-core'
 import { u8aToHex } from '@hoprnet/hopr-utils'
 import { AbstractCommand } from './abstractCommand'
-import { styleValue } from '../utils'
+import { styleValue } from './utils'
 
 export default class PrintAddress extends AbstractCommand {
   constructor(public node: Hopr<HoprCoreConnector>) {
