+++ packages/hoprd/src/commands/redeemTickets.ts
@@ -2,7 +2,7 @@ import type HoprCoreConnector from '@hoprnet/hopr-core-connector-interface'
 import type { Types } from '@hoprnet/hopr-core-connector-interface'
 import type Hopr from '@hoprnet/hopr-core'
 import { moveDecimalPoint } from '@hoprnet/hopr-utils'
-import { countSignedTickets, styleValue, toSignedTickets } from '../utils'
+import { countSignedTickets, styleValue, toSignedTickets } from './utils'
 import { AbstractCommand } from './abstractCommand'
 
 export default class RedeemTickets extends AbstractCommand {
