+++ packages/hoprd/src/commands/strategy.ts
