+++ packages/hoprd/src/commands/traverseChannels.ts
