+++ packages/hoprd/src/commands/utils/checkPeerId.ts
@@ -1,6 +1,6 @@
 import type HoprCoreConnector from '@hoprnet/hopr-core-connector-interface'
 import type Hopr from '@hoprnet/hopr-core'
-import type { GlobalState } from '../commands/abstractCommand'
+import type { GlobalState } from '../abstractCommand'
 import PeerId from 'peer-id'
 // @ts-ignore
 import Multihash from 'multihashes'
