+++ packages/hoprd/src/commands/utils/index.ts
@@ -3,7 +3,6 @@ export * from './displayHelp'
 export * from './isBootstrapNode'
 export * from './message'
 export * from './openChannels'
-export * from './parseOptions'
 export * from './cliOptions'
 export * from './knownConnectors'
 export * from './tickets'
