+++ packages/hoprd/src/commands/utils/isBootstrapNode.ts
