+++ packages/hoprd/src/commands/utils/knownConnectors.ts
