+++ packages/hoprd/src/commands/utils/message.ts
