+++ packages/hoprd/src/commands/utils/openChannels.ts
