+++ packages/hoprd/src/commands/utils/tickets.ts
