+++ packages/hoprd/src/commands/version.ts
